Standardize Nearby Station Access naming, refactor CRUD
